% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prep_data.R
\name{prep_data}
\alias{prep_data}
\title{Internal Ecospace Functions.}
\usage{
prep_data(ecospace, Smax)
}
\arguments{
\item{ecospace}{An ecospace framework (functional trait space) of class
\code{ecospace}.}

\item{Smax}{Maximum number of species (or other taxa) to include in
  simulation.}
}
\value{
Returns an emptry data frame with pre-defined number of species and
  functional traits of proper class type.
}
\description{
Internal functions not intended to be called directly by users.
}
\details{
Pre-allocate the basic data frame structure so classes (and factor
  levels, if used) are properly set. Bit clunky but efficient solution:
  adding first placeholder column allows \code{cbind} to work properly with
  data frame later, but requires deleting that placeholder at end.
}
\seealso{
\code{\link{create_ecospace}} for how to create an ecospace
  framework (functional trait space).
}

