% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample2.R
\name{sample2}
\alias{sample2}
\title{Internal Ecospace Functions.}
\usage{
sample2(x, ...)
}
\arguments{
\item{x}{Either a vector of one or more elements from which to choose, or a positive integer.}

\item{...}{arguments for particular methods.}
}
\description{
Internal functions not intended to be called directly by users.
}
\details{
Corrects behavior when object sampled has unit length. This is the same
function proposed in help file of \code{sample} as \code{resample}.
}
\seealso{
\code{\link[base]{sample}}
}
